requirements.txt
asgiref==3.8.1
black==24.8.0
build==1.2.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1
django-admin-honeypot-updated-2021==1.2.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cloudinary-storage==0.3.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-ipware==7.0.1
django-phonenumber-field==8.0.0
django-resized==1.0.2
django-unfold==0.37.0
djhtml==3.0.6
filelock==3.15.4
gunicorn==23.0.0
identify==2.6.0
idna==3.8
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
phonenumberslite==8.13.44
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.2.2
pre-commit==3.8.0
progressbar2==4.5.0
psycopg2-binary==2.9.6
pyproject_hooks==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
python-ipware==3.0.0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
virtualenv==20.26.3
whitenoise==6.7.0
